Directly import EvaluationStatus
