+++ src/lib/utils/synapseTypes/Evaluation/Evaluation.ts
@@ -1,11 +1,5 @@
+import { EvaluationStatus } from './EvaluationStatus'
 import { SubmissionQuota } from './SubmissionQuota'
-
-export enum EvaluationStatus {
-  PLANNED = 'PLANNED',
-  OPEN = 'OPEN',
-  CLOSED = 'CLOSED',
-  COMPLETED = 'COMPLETED',
-}
 
 // https://docs.synapse.org/rest/org/sagebionetworks/evaluation/model/Evaluation.html
 export type Evaluation = {
