+++ src/lib/utils/synapseTypes/Evaluation/EvaluationStatus.ts
@@ -1,0 +1,7 @@
+// https://docs.synapse.org/rest/org/sagebionetworks/evaluation/model/Evaluation.html
+export enum EvaluationStatus {
+  PLANNED = 'PLANNED',
+  OPEN = 'OPEN',
+  CLOSED = 'CLOSED',
+  COMPLETED = 'COMPLETED',
+}
